fix: update selection coord
